suppress default submit, read task description
have the task reading into memory however for some reason unsure how to actually append this task description into a <li> inserted into the correct location. am approaching this in a different  way to the lecturer code example because I am not storing this data in a separate file. will check previous labs on modifying elements